add user to the session and added add room page front end and back end

+++ src/main/java/com/yanni/sotrav/models/Room.java
@@ -30,13 +30,13 @@ public class Room {
 	private Date update_dt;
 	
 	@Column(name = "user_id", nullable = false)
-	private Integer user_id;
+	private Long user_id;
 
-	public Integer getUser_id() {
+	public Long getUser_id() {
 		return user_id;
 	}
 
-	public void setUser_id(Integer user_id) {
+	public void setUser_id(Long user_id) {
 		this.user_id = user_id;
 	}
